Move rich text module out of api modules
