+++ src/api/index.js
@@ -10,7 +10,5 @@ export * from 'src/api/calls';
 export * from 'src/api/notificationSettings';
 export * from 'src/api/callNotes';
 export * from 'src/api/messages';
-export * from 'src/api/richText';
 export * from 'src/api/imageUrl';
-export { default as richText } from 'src/api/richText';
 export { default as imageUrl } from 'src/api/imageUrl';
