+++ src/components/RichText/index.js
@@ -6,7 +6,7 @@ import SimpleMarkdown from 'simple-markdown';
 import mdStyles from 'react-native-simple-markdown/styles';
 import mdRules from 'react-native-simple-markdown/rules';
 
-import { RichText as RichTextObject } from 'src/api';
+import { RichText as RichTextObject } from 'src/richText';
 import defineTypes from './defineTypes';
 import baseRules from './rules';
 import baseStyles from './styles';
