+++ src/richText.js
